Publicize: Allow clearing Publicize custom message if a post title has been set (#18360)
When a post has a title added, the Publicize "Customize your message"
input will autofill with that post title if a custom sharing message has
not already been set.

Currently, if a post has a title added, and users select and then try to
delete (not type over) all of the text in the custom sharing message
input, it will immediately fill with the post title text.

This change will track if the custom message input has been editied, and
if it has, allows a user to select and delete all text, even when the
post has a title added to it.\n\nCommitted via a GitHub action: https://github.com/automattic/jetpack/runs/496456180