GH: Rewrite required-review action to add path-based requirements (#18561)
As we start pulling more different things into the monorepo, the old
"Must have a Crew Review" check isn't going to work for us.

So rewrite the check to allow for different team requirements by paths
touched in the change.

Also, start running the workflow on opened and updated PRs too. GitHub
sometimes keeps existing reviews when pushing a merge commit, but
drops status checks so we need to re-apply it.

Co-authored-by: Brandon Kraft <public@brandonkraft.com>\n\nCommitted via a GitHub action: https://github.com/automattic/jetpack/runs/516427979