Build: Fix more always-changing stuff (#18579)
* GH: Use consistent build directory for build workflow

Use a constant build directory in the action because the path winds up
in vendor/composer/installed.json for all the monorepo packages.

* Set COMPOSER_ROOT_VERSION in build-production-php

Without this (or .version or .extra.branch-version in composer.json),
composer pulls the git commit sha for a 'reference' field in
vendor/composer/InstalledVersions.php. And unlike when it does so for
monorepo packages, it doesn't even check for the most recent commit that
actually touched projects/plugins/jetpack/.\n\nCommitted via a GitHub action: https://github.com/automattic/jetpack/runs/518516659