Move Whatsapp styles to Send-a-Message directory (#18947)
The block naming convention required to enable plan checks on the
WhatsApp block is not compatible with the way we load styles in
the render callbacks. We can get around this by loading the
styles from the parent block.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/602629422