Add button width options for Jetpack Button and Subscriptions Form (#18809)
* [not verified] First pass add width selector to Button

* [not verified] Add button width on frontend

* [not verified] Override alignment CSS to preserve custom widths

* [not verified] Fix button width in contact form when alignment selected

* [not verified] Load styles for frontend

* [not verified] Add non-percentage width controls and disallow percentage width when floated

Taking code from https://github.com/Automattic/jetpack/pull/18872

This adds in an additional width control for setting custom width by %/px/em,
and disables the preset width options and custom % options if a left or
right alignment has been selected on the button (to circumvent the issues
with percentage width on floating divs).

* [not verified] Persist button width on save

* [not verified] Add button width to Subscriptions form

* [not verified] Remove 100% width on contact form button

* [not verified] Add maximum width to contact form button

* [not verified] Account for margins in width of Subscription form button

* [not verified] Update the unit when a percent width preset is selected

The unit does not update automatically because updating the value
prop does not cause the onChange event to fire on the UnitControl.
We get around this by keeping track of the unit at the parent
level until this bug can be fixed.

* [not verified] Subscription form buttons on new line adhere to max width

* [not verified] Reset unit as well as width when button is floated

* [not verified] Set max to 100 when width contains %

* [not verified] Add changelog.

Co-authored-by: Jeremy Herve <jeremy@jeremy.hu>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/647065869