build: Sort by dependencies (#19094)
Currently we rely on some manual ordering of the projects and some luck
that projects that actually have a build step come alphabetically before
the projects that depend on them.

What we really should be doing is ordering the build by dependency, to
guarantee that each project's dependencies are build before the project
itself.

This PR does that, splitting the "determine dependencies" code out from
list-changed-projects.php and creating a new get-build-order.php that
outputs the project slugs in an order they can be safely built.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/661688268