Calypsoify: Fix use_block_editor_for_post_type undefined error (#19204)
* Calypsoify: Fix use_block_editor_for_post_type undefined error

In #17826 I added the check for whether to load the block editor in the
initial setup function, which gets called during `wp_loaded`. If the
editor was being loaded for a post type other than a 'post' then a call
was made to `use_block_editor_for_post_type`, which wasn't yet
available.

This change moves all the setup and initialisation to the `admin_init`
hook, which is late enough for the file with this function to have been
included. It seems there are no checks for `is_calypsoified_enabled`
between these two points in the life cycle, so this should be safe.

* Adding the changelog file

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/662292381