Publicize: update Twitter text processing library (#19345)
* Publicize: update Twitter text processing library

The lib was updated to update a PHP 8 fix for a part of the lib we do not use in Jetpack.
I do not think it hurts to update in any case.

Related PR: https://github.com/nojimage/twitter-text-php/pull/39

* Add changelog

* Update projects/plugins/jetpack/changelog/update-twitter-text-php-213

Co-authored-by: Brad Jorsch <anomiex@users.noreply.github.com>

Co-authored-by: Brandon Kraft <public@brandonkraft.com>
Co-authored-by: Brad Jorsch <anomiex@users.noreply.github.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/702937314