Site Editor Compatibility: Podcast Player block issues in editor and front end views (#19578)
* [not verified] For TT1 block-based theme at least, we have to wait until the dom is loaded before we execute DOM-based queries.

* [not verified] # This is a combination of 3 commits.

I don't like doing this, but there's no other way to enqueue/inject the styles we need for the podcast player into the site editor iframe.
So, using JS and the DOM, we find the CSS we need from the head of the parent window and move it into the site editor.
YAY.

Changelog

Adding null checks to the elements

we should ensure that the parent-level dom elements are there before we use them

* Using the block element to check whether it's sitting inside an iframe now before we load.
Abstracting the element copier to be more generic.

* Remove unused method

* Switch from useEffect to ref callback for getting podcast player element ref

* Making element removal optional.

In the case of style sheets it won't matter much for our current use case (podcast player) but for other elements we might want to leave the element in both contexts. I can't for the life of me think of a reason why right now.

* This updates the iframe check to also look for the block editor iframe name.
It's not 100% bullet-proof, but aligns with the general theme of ephemerality in relation to these site editor bug fixes

* Returning an array from maybeCopyElementsToSiteEditorContext and not mixed return types
Testing whether we indeed have access to the parent DOM before we try to avoid fatals

Co-authored-by: Glen Davies <glen.davies@a8c.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/784428813