Button block: check block root parent attributes (#19883)
* When passing through attributes to a child Button block using `passthroughAttributes`, we want to make sure the Button looks to the correct parent block to fetch the attributes.
If a Calendly block is nested inside another block, previously the Button was looking to the top of the nesting tree rather than it's parent.

So, rather than use `getBlockHierarchyRootClientId`, looks at the top of the tree in which a block is nested, we look to `getBlockRootClientId` so that we can get the immediate parent block.
Updating README to reflect this information.

* changelog

* Update fix-passthrough-attributes-get-parent-attributes

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/861667173