Make carousel more robust against missing data (#20101)
The carousel currently fails if some crucial data properties are missing
in the images it attempts to load.

This will make it more robust, allowing it to fall back to other ones.
In addition, it restores the pointer cursor in some situations.

Fixes #11191.

Co-authored-by: Brandon Kraft <public@brandonkraft.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/957165023