Tiled gallery: Don't swap width and height in srcset (#20424)
It had the width and height swapped when processing images to generate
the srcset attribute. This is most noticeable when the gallery contains
images that are much wider than they are high, resulting in ridiculously
small images in the srcset that the browser then scales up into a blurry
mess.

Fixes Automattic/jpop-issues#6772

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1053038870