tools: Add config for the --type for tools adding change entries (#20511)
In a few places we have it special cased to do `--type=other` instead of
`--type=changed` when auto-adding a change entry for plugins/jetpack.

Instead, let's make that a configuration in composer.json so if we have
other plugins that copy Jetpack in the future then we don't have to find
all the scripts with special casing.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1147687159