actions: Don't run Beta Plugin artifact action on forked PRs (#21019)
The existing check that was meant to avoid it was incorrect.
`github.repository` is the repo the PR is created on, not the source
repo, so checking it is mostly pointless. We need to look at
`github.event.pull_request.head.repo.full_name` instead.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1222070157