jetpack: Replace unnecessary `page` dep (#21122)
We haven't been using this library correctly in the first place, and it
has now broken the incorrect way we were using it.

It seems that the way we were using it works out as equivalent to a
single line of JS, so just do that.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1257900399