Jetpack: add a fallback for is_connected in the is_connection_ready method (#21210)
If an older version of the Connection package without the `Manager::is_connected' method
is loaded and Jetpack is loaded in a way that can't be detected ahead of time (for example
via CLI), a fatal error can occur during Jetpack activation.

To fix this, add a fallback to the 'Manager::is_registered' method which is available in
older versions of the Connection package.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1282849525