Use webpack config JS package in Jetpack (#21581)
This shouldn't change anything in this iteration, but double-checking
might be a good idea.

This does take the opportunity to fix the extension build to keep
plugins and blocks in the order defined in extensions/index.json,
where #20994 had put all the blocks before all the plugins.

This also removes uses of calyspo-build in Jetpack's jest config.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1435311670