jetpack: Don't set Webpack's `output.pathinfo` in production builds (#21727)
There's no point to it, the generated comments just bloat the
CSS and license.txt files.

Note Webpack already defaults it to true for `mode: development` builds,
where it makes a little more sense for the comments to be present.

It's not clear why it was ever being set in the first place,
neither #12463 nor #13070 seem to have any discussion about the
additions.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1453633526