webpack-config: Fork calypso-build's mini-css-with-rtl plugin (#21595)
And clean it up with some better formatting, better input validation,
and a better way to override the default "isRTL" expression.

This removes our last dependency on calypso-build that's not going away
for other reasons in other PRs, which will save us a bunch of indirect
deps that tend to be out of sync.

Also, split the "RtlCssPlugins" into two now that StandardPlugins looks
to be the most common way of getting plugins. That makes it more
straightforward to configure the two plugins.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1454128643