Publicize: Create new objects in the store effect (#21663)
Previously we would copy over the any existing object based on the
connection ID, which meant that if any data had been updated (like the
profile image) it wouldn't be seen as a change, because the object
reference would still be the same.

This creates a new object for each connection in response to the API
results. It will force a rerender, but I think that would happen anyway
as we dispatch the action to update the meta data already.

We could potentially check for any of the properties having changed
instead. I'm not sure if this is a necessary optimisation though.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1476714083