Disable CSSTidy property shorthand optimization in the editor or headstart (#21891)
* Disable CSSTidy property shorthand optimization in the editor or headstart

Summary:
This fixes block validation errors, shown in the editor on a blank page load and also fixes block validation issues associated with the `pub/payton` homepage starter content.

Building on the work in D69931-code and D70357-code

The gutenberg editor does not support property shorthands such as `padding: 10px 20px;` instead, the property must be specified in long form `padding-top:10px; padding-right:20px;` etc.

The CSSTidy processor is used to optimise css both when page content is fetched, and when headstart is run. We are currently questioning whether it is still a useful part of wordpress.com https://[private link]

We override CSSTidy's `filter_attr` function and there is already some code with the intention of disabling the shorthand optimization in the gutenberg editor. But it only checks whether the `_gutenberg_nonce` is present. It appears that other requests are made when the editor is loaded that do not contain the nonce, but that will still be validated by gutenberg. I haven't determined the source of these requests, but I believe that they are the requests to load block patterns for the editor. In any case, they are returning shorthand properties to the block editor which, as a result, is logging validation errors to the console.

This investigation was originally started to debug validation errors with the `pub/payton` theme. p1635985973066500-slack-CBTN58FTJ. Disabling the shorthand property optimization for headstart prevents the block validation issues specific to the `pub/payton` theme.

Reviewers: arthur791004, philipmjackson, ramonopoly, fullofcaffeine

Reviewed By: arthur791004, philipmjackson, fullofcaffeine

Subscribers: fullofcaffeine, simison, philipmjackson, arthur791004, ramonopoly

Tags: #touches_jetpack_files

Differential Revision: D70425-code

This commit syncs r235710-wpcom.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1517051300