Subscription Block: Add success message  (#21273)
* [not verified] Blog Subscription Widget: Added transform to block

* [not verified] Remove widget from Legacy block

* Subscription Block: Implement success message for subscription block.

In this commit, I have updated all the files of subscription block extension
I have also updated the modules views.php file to support success_message as a
shortcode attribute.

* Added label and wrapped string with i18n function

* fixing linting error for successMessage default value in attribute.js

* Add support for deprecated version v6. Fixed test cases.

In this commit, I have reverted few of the commits since they're not correct and
addressed review comments. Also introduced new deprecation version v6.
All test cases are passing.

* update deprecation version reason comment

* update changelog file, write correct changelog message

This PR is created from other branch which was under dev(that time, now it's merged) so changelog message should be related to this PR

Co-authored-by: Samiff <samiff@users.noreply.github.com>

* Update intialOpen props defaults to false

Co-authored-by: Samiff <samiff@users.noreply.github.com>

* Remove successMessage from deprecated v6 save.js

* Add descriptive comment of new block attribute

It is always better to write descriptive comment so that it will be easier to understand in future or by someone who looks into it.

Co-authored-by: Samiff <samiff@users.noreply.github.com>

* Address review comments

- Add attributes.js file in deprecation v6
- Remove migrate block in deprecation v6
- Fix test cases

* Update import attributes in deprecation v6 save.js

Co-authored-by: Samiff <samiff@users.noreply.github.com>

* fix save.js in v6

Co-authored-by: Samiff <samiff@users.noreply.github.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1551150310