Add babel-plugin-replace-textdomain (#21943)
This is a (surprisingly simple) Babel plugin intended for use when
building Webpack bundles using Jetpack RNA components, so the resulting
bundles will use the plugin's textdomain (since that's where WordPress's
translation infrastructure will put the strings) instead of each
components'.

This also adds it as an option in webpack-config, but it's not enabled by
default as it has required configuration.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1564895942