Assets: Add textdomain aliasing (#22083)
This adds the ability to register textdomain aliases with the Assets
package.

Assets then hooks into `__()` and so on to have them check the
aliased-to domain if no translation is found for the aliased-from
domain. It also hooks into script translation file loading to load a
translation file for the aliased-to domain if no file exists for the
aliased-from domain.

This brings us one step closer to fixing #21690. Followups will switch
all the packages to use alias-able textdomains (instead of them all
using "jetpack") and arrange for the aliases to actually be registered.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1582970761