Update minimum pnpm version (#22110)
We need a version at least 6.14.6, which fixes a bug in handling
overrides. But let's not go all the way to 6.24 just yet, that adds a
long output of missing peer deps that we're not ready to deal with just
yet.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1584948169