Jetpack: activate default modules when the site is already connected (#21353)
* Jetpack: activate default modules when the site is already connected

Currently, the default Jetpack modules are activated when the site is registered and when the
first user is authorized. However, there are other plugins that can be used to register the site
and authorize the user. If the site is already connected via another plugin when the Jetpack
plugin is activated, the default modules will not be activated.

Fix this by adding a new method, `handle_default_moduule_activation`, and use it to activate the
default modules during plugin activation when the site is already connected and when
the site is registered.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1654540313