[not verified] Jetpack modules: Fix uses of `this` for `window` (#22226)
Prior to #21808, the JS files were only minified, not bundled. That
preserved use of `this` at the file level to mean `window`.

That PR changed the build to use Webpack, as a start towards things like
using npm modules instead of copy-pasting libraries into the code. But
when bundled, `this` at the file level is no longer `window`. So these
uses need to be changed to use `window` directly.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1660407623