Admin: update upgrade links to use new instertials when possible (#22531)
* Admin: update upgrade links to use new instertials when possible

We've introduced new in-admin screens to promote upgrades; let's use them instead of directly linking to the checkout screen in Calypso, when possible. It's worth noting that for some of our upgrades (ads / Google Analytics), we do not have dedicated screens yet, so we'll continue to link to Calypso for those.

* AUpdate VideoPress upgrade link too.

* Remove unneeded prop

It's no longer used since #21006.

* Update PlanIcon.plan prop type

Co-authored-by: Samiff <samiff@users.noreply.github.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1758404501