Premium content: Improve editing consistency (#22626)
* Premium content: Improve editing consistency

In the editor, viewing the content for people who don't have a
subscription is sometimes referred to as 'Visitor view' and sometimes as
'Logged out view'. This change makes the interface more consistent by
using 'Guest View'.

It also swaps the order of the added content (subscriber/guest view) to
match the order of the menu buttons, so that the Guest view is the first
content inserted and then the Subscriber view.

Fixes #22425

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1795337494