Twitter Gather: make endpoint restricted to contributors (#22657)
Until now, the endpoint was available to anyone, and one could pass any Twitter Thread to that endpoint to make a blog-authenticated request to WordPress.com, and then, in turn, an authenticated request to the Twitter API using our app credentials.

Here is an example:
https://jeremy.hu/wp-json/wpcom/v2/tweetstorm/gather?url=https://twitter.com/KieranBinchy/status/1488812175372267526

That endpoint is currently used in the block editor (so only by folks with contributing capabilities on a given site) to "unroll" a Twitter Thread, i.e. pull all tweets from a thread based off a single tweet.

I think it would be fine to limit that endpoint to contributors.

- Related post: p3btAN-1Fc-p2
- This merges r239582-wpcom from WordPress.com (D74346-code).

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1796778811