[not verified] Related Posts: Remove duplicated filter (#23225)
In #22270 we added a second call to the
`jetpack_relatedposts_returned_results`, so the output could be
modified, but it was passed a single related post rather than an array
of results.

Because of this, any existing code hooked into the filter caused a fatal
error as it tried to process a scalar as an array.

This PR reverts the change. It looks like the output can still be
modified in the same way as the testing instructions for #22270 without
it.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1931860756