Google Fonts: update the method used to preconnect Fonts source domain. (#23322)
* Google Fonts: update method used to preconnect Fonts source domain

Follow-up to #23045

Let's favor using a Core implementation with a Core filter instead of manually adding to the head.

* Use filter instead of action

* Only add resource hint if the site supports the Webfonts API

See #23306

* Update readme with latest method name.

* Introduce new class that doesn't depend on WP_Webfonts_Provider

WP_Webfonts_Provider may not always be available, so we cannot call a method from a child class of `WP_Webfonts_Provider` from a WP filter. Let's introduce a new class instead, that doesn't depend on anything, and do the class check inside there. That should simplify things a bit for the folks having to implement that filtering.

Co-authored-by: Brandon Kraft <public@brandonkraft.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1983275068