Slideshow Shortcode: Adding the Jetpack namespace to CSS class names (#22675)
* Adding the Jetpack namespace to slideshow shortcode CSS class names.

This is to prevent CSS name collisions. This involves changes to the CSS file as well as the PHP and JS files where the class names are first defined.

* Adding the Jetpack namespace to slideshow shortcode CSS classes mentioned in the test file.

* Adding the Jetpack namespace to slideshow shortcode CSS class names mentioned in the theme compat files

This affects only the slideshow-window CSS class names mentioned in the test file.

Co-authored-by: Samiff <samiff@users.noreply.github.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/2053850531