tools: Add mirror repo check to changelogger-release.sh (#25132)
Two weeks in a row now we've run into problems because we run
changelogger-release.sh, then try to run the release branch script and
it breaks because a package is missing a mirror repo.

Let's avoid that in the future by having changelogger-release.sh fail if
a package is missing the mirror repo config. But don't fail if we're
"releasing" the package directly. And don't fail on js-packages, as
those currently have different rules (and mostly lack mirrors).

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/2705005996