Update credentials setup message (#26254)
* Update credentials setup message

We're updating the message shown when credentials are not set up in
Calypso blue with the same with we have in Calypso green, for feature
parity and to avoid customers contacting support because they think
credentials are needed for backups to work.

* Update credentials message for consistenty

Updating the message around server credentials in two more places for
consistency.

* Fix typo

* Update credentials message for the Scan section

Yet another place that needed a copy change for consistency reasons.
Credentials are needed to enable one click fixes for Scan.

* Replace one click with one-click in credentials texts

We're already using the hyphenated word in support documentations and
other places.

* Fix linting error

* Update wording on credentials setup messages

We're trying to improve clarity and increase consistency with other
places where we show a credentials setup banner.

Co-authored-by: nunyvega <alvaro.vega@automattic.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3133655576