VideoPress: Check connection before getting connected-only data (#26685)
* Check site connection before requesting connected-only data

We are moving all the initial state data request that may need an active connection to it's own method, wrapped around a connection check, to prevent requesting it too soon or when the data will not be available.

Currently the only data in this situation is the site info related to storage usage.

* Add changelog file

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3205126267