Jetpack: Replace remaining `jetpack_require_lib()` calls (#27094)
Mostly these are things which only exist in Jetpack or only exist in
wpcom, so there was no de-Fusioning to be done. Any in Jetpack are
replaced with a direct `require_once`, and any in wpcom are changed to
use wpcom's `require_lib()` function instead.

Two of note:

* There's a non-Fusioned "widgets" lib on both, with the Jetpack version
  defining a class named "Jetpack_Widgets". The two uses here both used
  the latter class.
* "class.media" was de-Fusioned in D90466-code, I just forgot to ever
  follow up.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3345963724