VideoPress: New Control in VideoPress Block to Add VideoChapters Block (#27240)
* Add: Video chapter block prompt when VTT present

Updated the VideoPress editor block to show a prompt to guide the user
to add the new Video Chapters block after the user adds chapter links to
the block description.

* changelog

* Modify: VideoChapter prompt styling in VideoPress

Improved the styling of the prompt to add a VideoChapter block when
editing a VideoPress block.

* Fix: VideoChapters prompt dismiss

The prompt in the VideoPress block to add a VideoChapters block is now
dismissable.

* tools/fixup-project-versions.sh

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3381640994