VideoPress: Use playback tokens to get resources when the VideoPress site privacy is set to private (#27380)
* Provide videoIsPrivate flag on the client initial state

Also, use the flag as the reference to decide if the poster should be null on the first load, preventing the load of resources that will 403 while we don't fetch the required playback token.

* Add support to the new video_is_private response field on the VideoPress video types

The field is called videoIsPrivate on the types.

* Map the new field from the response object to the video object

* Use the new videoIsPrivate flag to decide if the video needs a playback token

* Change list component to use the new videoIsPrivate as the source of the privacy state

* Return list component privacy source to original value so it reflects the actual setting of the video

This way, despite the video being private when the site is private as well, the user see the actual setting on the video list.

* Add changelog file

* Rename the videoIsPrivate flag to a more contextual needsPlaybackToken

This way, it's value gets tied to the idea of the playback token need, that is the usage that we were making for the flag anyway.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3464666335