Change privacy filters to include the site default setting when appropriate (#27603)
When the site default privacy setting is private, and the search is looking for private videos, the videos with site default privacy setting are included.

The same way, when the site default privacy setting is public, and the search is looking for public videos, the videos with site default privacy setting are include.

This way, the videos are listed based not only on the privacy setting itself, but also taking into account the site default setting as well.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3548510361