Jetpack: Remove dead static-site-generator-webpack-plugin dep (#27889)
The dependency seems to be dead, and has some pretty outdated
dependencies. Let's copy the code into the monorepo, and clean it up to
remove the features (and corresponding deps) we don't use.

Even better would be to remove the need for it entirely, but
[work on that seems to have stalled out][1].

[1]: https://github.com/Automattic/jetpack/pull/23214

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3680193845