Blocks: iterate on the design of the Promote post-publish panel (#27960)
* Blocks: add a new Analytics hook

This should be used in the Jetpack plugin instead of the currently hardcoded analytics pulled from the React dashboard.
We can then leverage it in other plugins as well.

* Fix imports

* Get connection data from conn. package instead of Jetpack data

This should prove easier to use by other standalone plugins, which do not have the same data passed to JavaScript via Jetpack_Editor_Initial_State.
Grabbing the data from the connection package should consequently be easier.

* Expand docs

* Blocks: switch to new shared, extracted implementation of Tracks

This is a follow-up to #27937.

* Blocks: update Promote button in post-publish panel

Let's turn the simple link into a button so it is more visible. This also follows the pattern already in use in the Anchor post-publish panel.

This also updates tracking for this button click to use the new tool introduced in #27937

* Update the icon

See https://github.com/Automattic/jetpack/pull/27960#issuecomment-1362793951

* Rename tracks event

Co-authored-by: Derek Smart <dsmart@dsmart-mbp.lan>
Co-authored-by: Derek Smart <smart@automattic.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3766163898