Jetpack: Masterbar: Return correct type for get_available_languages filter (#28622)
Code in Masterbar hooks the get_available_languages filter on WoA sites
to return an empty value to disable a language dropdown. `__return_null`
was used to do this, which works but can make other code that tries to
actually use the value warn (in PHP <8) or fail (in 8.0+) since it's
expecting an array.

Change it to `__return_empty_array` which will return the correct type.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4027658293