[RNMobile] Register VideoPress block (#28812)
* Register VideoPress block

* Add changelog to VideoPress package

* Add placeholder functionality for VideoPress block

With this commit, native-specific edit.js and style.scss files are added with placeholder components. This serves as a starting point for the VideoPress block.

* VideoPress: Register attributes

* Add native-specific index file

The native-specific index file is necessary in order to limit the block in Mobile Gutenberg.

* Check 'available' flag before registering block

* Remove unused "example" setting

* Move conditional to constant for conciseness

* Replace VideoPlayer with temporary placeholder

---------

Co-authored-by: Renato Augusto Gama dos Santos <renato_0603@hotmail.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4178601983