VideoPress Package: Refactor useSyncMedia hook (#29898)
Refactor part of preparation for implementing a native version of the useSyncMedia hook. Note, this will be the first of a series of planned PRs for splitting the hooks functionality into smaller hooks, with the goal of making it more straightforward to implement the native functionality.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4610546769