[not verified] Stats: display links to post's stats even when not connected (#31025)
With Odyssey stats, one can now access stats in wp-admin, even when not connected to WordPress.com, as long as their role gives them access to stats. We can consequently display the icon in the posts list as soon as the user is allowed access to stats.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5117290197