Jetpack: Fix `plugins/new` and `themes/new` endpoints (#31090)
These were declared as taking a file upload, but in reality they depend
on wpcom code intercepting the upload, saving it as an attachment, and
then forwarding it on to the site. Fix the declaration to reflect that.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5125637120