Tock Block: Implement settings with a dropdown (#31355)
* Tock Block: Update toolbar and popover

We decided to move the settings popover to the toolbar, instead of
underneath the block, as well as change the button to text rather than
an icon.

This also includes a couple of grammatical tweaks to the helper and
description text.

* Upade the onClose logic.

* Reset the input when the block loses focus.

* Change the block to production

* Update the styling based on feedback.

* Adjust example render width.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5265358174