Publicize Components: Refactor the media validation (#31565)
* Publicize Components: Refactor the media validation

We need to change the way we validate the media to be shared. Previously
we checked the lowest requirements for all enabled connections, but
given the way that Instagram requires an image we were forcing the media
to be valid for Instagram, which is currently only a JPG.

This change reworks the validation so we do it on a per connection basis
and disable the individual connections if the selected media is invalid.

* [not verified] Update the notice message when there is validation errors.

* [not verified] Fix the typo, and stop calling Object.keys on the validationErrors.

* Restore the correct Facebook image size requirement.

* Fix media restricitons tests

* Update package version

* Revert to not adjusting the the connection state.

---------

Co-authored-by: Gergely Juhasz <gergely.juhasz@automattic.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5402721139