Jetpack Custom-CSS: Do not `@import` from the filesystem (#33959)
The SCSS and LESS libraries we're using by default can attempt to
process scss or less files from the filesystem for `@import` directives.
There shouldn't be any need for this, so to be safe let's turn it off.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/6797193183