WordAds: ensure the CMP banner works in production environments (#35791)
* Cookie Consent: document new filter usage

Follow-up to #35165

Adding the docblock will help the codex generator create the necessary codex page for the new hook.

* CMP Loader: force the function to be used

The function was dropped from the production build since it wasn't used within the file. This workaround ensures the production build includes the function, and thus isn't empty.

* Switch to Asset method to enqueue the file and fix file reference

The production file we tried to load wasn't correct. This commit fixes that, and switches over to the `register_script` method which is a bit more readable.

* Simplify enqueuing of the second resource

We do not need to duplicate the URL, since it's the same regardless of the environment.

* Simplify building endpoint URL and sanitize URL parts

* Add changelog

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/7986649208