Backport changes for jetpack-image-cdn@0.4.8 (#39213)
* image-cdn: Avoid fatal on bad img width/height (#39208)

If the `<img>` tag has a value for `width` or `height` that's neither an
integer nor a percentage, this will cause a fatal error when the value
is attempted to be used as an integer. Add validation to avoid this.

* Update package version

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/10689684479

Upstream-Ref: Automattic/jetpack@f43a0b43d50da62e1f71649c31e1237202cc69d9