eslint: Replace abandoned eslint-import-resolver-exports (#40385)
The maintainer of this package has abandoned it in favor of
eslint-import-resolver-typescript. So let's switch to that.

The new resolver seems to do a better of a job of finding files for some
of the more unusual cases, which lets the `import/order` rule detect
things as "internal" that were previously defaulting to "external",
which changes the order it enforces.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/12124747647

Upstream-Ref: Automattic/jetpack@9742576c494a14a4f53ee2418f83eaa72eadd12e