Social: Handle connections with the must_reauth status (#41494)
LinkedIn connections need to be reauthenticated every year. Currently
that is every 60 days, but as we release the change to use the refresh
token, it will mean that all LinkedIn connections will be marked as
needing to be reauthenticated.

The connections will still be able to be used, but the UI will now say
that they need reconnecting and follow a similar process to the broken
connections.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/13114353989

Upstream-Ref: Automattic/jetpack@99459eb0e1b45f1615c361f6ab9fd7ceca9d32f7