Related Posts: ensure authors+ have access to the block (#42967)
Fixes #42966

By opening up this endpoint to authors, we'll be able to check that the related posts feature is available, even for folks who are not admins.

This of course means authors will have access to related posts options. I think this is okay, the options that can be returned are not necessarily problematic. For reference, this is information that could be visible:

- enabled
- show_headline
- headline
- show_thumbnails
- show_date
- show_context
- layout

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/14570471739

Upstream-Ref: Automattic/jetpack@22f069528a3af11099d4b1f44038c2a47fb15b90