Forms: Add unread/read filtering (#45514)
* Update the endpoint to allow filtering by unread

* Update the counts endpoint to to return read and unread state

* Update the front end to show the filter and update the count.

* Update the count to always include a the current Query as key so that the filters always return the expected results.

* changelog

* Update the empty state to account for the filter.

* Invalidet the count store when the user marks things as unread so that the count is more accurate

* normalize values in cache key

* Invaliade even if only one of the requests succeeded

* Fix wording.

Co-authored-by: Miguel Lezama <miguel@automattic.com>

* don't double sanitize

* Fix: Keep track of the fields that are marked as read and need to be refetched.

* Make the js tests pass

* Fix pagination bug.

Currently when you are using the unread/read filter and you mark things as read the pagination breaks. Since the state on the client and the backend is not in sync any more. We fix this by passing in an array of invalid ids. This way the state between the browser and the backend matched and the pagination works as expected.

* Fix phan

* fix js tests

---------

Co-authored-by: Miguel Lezama <miguel@automattic.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/18604924152

Upstream-Ref: Automattic/jetpack@bf1d20a92546b3ffabe1ece4f1dac048739244e0